Send correct id for each flag
